# openapi.yaml
openapi: 3.0.0
info:
  title: Team Task Tracker API
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [ADMIN, MANAGER, MEMBER]
                organizationName:
                  type: string
      responses:
        '201':
          description: User registered successfully
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
  /projects:
    get:
      summary: Get projects
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of projects
    post:
      summary: Create a project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Project created
  /tasks:
    get:
      summary: List tasks
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: assigneeId
          schema:
            type: string
      responses:
        '200':
          description: List of tasks
    post:
      summary: Create task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                projectId:
                  type: string
                assigneeId:
                  type: string
                priority:
                  type: string
                dueDate:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Task created
  /tasks/{id}:
    put:
      summary: Update task
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                status:
                  type: string
                assigneeId:
                  type: string
      responses:
        '200':
          description: Task updated
    delete:
      summary: Delete task
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task deleted